+++ codex/serializers/metadata.py
@@ -1,15 +1,8 @@
 """Codex Serializers for the metadata box."""
 from rest_framework.serializers import CharField, IntegerField
 
-from codex.serializers.mixins import (
-    BrowserAggregateSerializerMixin,
-    get_serializer_values_map,
-)
+from codex.serializers.mixins import BrowserAggregateSerializerMixin
 from codex.serializers.models import ComicSerializer
-
-METADATA_ORDERED_UNIONFIX_VALUES_MAP = get_serializer_values_map(
-    [BrowserAggregateSerializerMixin], True
-)
 
 
 class MetadataSerializer(BrowserAggregateSerializerMixin, ComicSerializer):
@@ -17,3 +10,5 @@ class MetadataSerializer(BrowserAggregateSerializerMixin, ComicSerializer):
 
     filename = CharField(read_only=True)
     parent_folder_pk = IntegerField(read_only=True, required=False)
+    series_volume_count = IntegerField(read_only=True)
+    volume_issue_count = IntegerField(read_only=True)