+++ codex/urls/api/v3.py
@@ -1,12 +1,8 @@
 """codex:api:v3 URL Configuration."""
-from django.urls import include, path, register_converter
+from django.urls import include, path
 from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView
 
-from codex.urls.converters import GroupConverter
 from codex.views.version import VersionView
-
-register_converter(GroupConverter, "group")
-
 
 app_name = "v3"
 urlpatterns = [