+++ codex/views/opds/v2/start.py
@@ -1,0 +1,28 @@
+"""OPDS 2 start view."""
+from django.http.response import HttpResponseRedirect
+from django.urls import reverse
+
+from codex.serializers.choices import DEFAULTS
+
+
+def opds_2_start_view(request):
+    """Redirect to start view, forwarding query strings and auth."""
+    kwargs = DEFAULTS["route"]
+    url = reverse("opds:v2:feed", kwargs=kwargs)
+
+    # Forward the query string.
+    path = request.get_full_path()
+    if path:
+        parts = path.split("?")
+        if len(parts) >= 2:  # noqa PLR2004
+            parts[0] = url
+            url = "?".join(parts)
+
+    response = HttpResponseRedirect(url)
+
+    # Forward authorization.
+    auth_header = request.META.get("HTTP_AUTHORIZATION")
+    if auth_header:
+        response["HTTP_AUTHORIZATION"] = auth_header
+
+    return response